fix: fix classes and add turn and win files
